refactor: extract item frame break related events to independent listener
